requirements.txt
hyperliquid-python-sdk>=0.10.1
requests>=2.31.0
websocket-client>=1.5.1
selenium>=4.18.1
webdriver_manager>=4.0.1
beautifulsoup4>=4.12.3 